Separate graph implementations into separate files
